Introduced new type of Hits rendering, namely Box
This allows to display hits as boxes. The syntax is identical to line, meaning 6 coordinates per hit. The first 3 coordinates define the center of the box, the 3 others are the sizes along the 3 main axis.
At this stage, the box cannot be oriented. Its edges are parallel to the main axis