Use full path for object names
In case of 2 objects with identical name "Mirror" living in 2 different menus "Rich1" and "Rich2", it was so far impossible to identify them uniquely, as both names were just "Mirror".
With this change, the names are "Rich1 > Mirror" and "Rich2 > Mirror". Menu display has been adapted so that it stays the same